crimson/zns: write segment tail at write pointer
SegmentAllocator::close_segment() writes tail information to a
segment before closing the segment, and this is written at the
end of segment. However, for ZNS SSDs, the writes have to always happen
at write pointer, so writing tail info at the end of a zone fails.

Added write_segment_tail() function which writes the tail information
at write pointer for ZNS device, for a regular device it continues to
write at the end of segment.

This would also require a read_segment_tail() to read the tail information
from correct location for ZNS SSDs.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>